requirements.txt
python-telegram-bot==13.11
python-dotenv==0.20.0
gspread==5.3.0
requests==2.27.1
